project/requirements.txt
requests
beautifulsoup4
openpyxl
PyYAML